requirements.txt
streamlit
validators
requests
python-dotenv
langchain
langchain-community
langchain-groq
youtube-transcript-api
unstructured
pypdf